move custom_hilbert common logic to HomogeneousHilbert
