+++ netket/hilbert/qubit.py
@@ -19,11 +19,11 @@ from jax import numpy as jnp
 
 from netket.graph import AbstractGraph
 
-from .custom_hilbert import CustomHilbert
+from .homogeneous import HomogeneousHilbert
 from ._deprecations import graph_to_N_depwarn
 
 
-class Qubit(CustomHilbert):
+class Qubit(HomogeneousHilbert):
     r"""Hilbert space obtained as tensor product of local qubit states."""
 
     def __init__(self, N: int = 1, graph: Optional[AbstractGraph] = None):
